+++ test/plausible_web/controllers/api/external_sites_controller_test.exs
@@ -384,4 +384,20 @@ defmodule PlausibleWeb.Api.ExternalSitesControllerTest do
              }
     end
   end
+
+  describe "GET /api/v1/sites/:site_id" do
+    setup :create_new_site
+
+    test "get a site by it's domain", %{conn: conn, site: site} do
+      conn = get(conn, "/api/v1/sites/" <> site.domain)
+
+      assert json_response(conn, 200) == %{"domain" => site.domain, "timezone" => site.timezone}
+    end
+
+    test "is 404 when site cannot be found", %{conn: conn} do
+      conn = get(conn, "/api/v1/sites/foobar.baz")
+
+      assert json_response(conn, 404) == %{"error" => "Site could not be found"}
+    end
+  end
 end