+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - New filter type: `contains`, available for `page`, `entry_page`, `exit_page`
 - Add filter for custom property
 - Add ability to import historical data from GA: plausible/analytics#1753
+- API route `GET /api/v1/sites/:site_id`
 
 ### Fixed
 - UI fix where multi-line text in pills would not be underlined properly on small screens.